AlertToCareTests/ControllerTests/MonitorControllerTest.cs
using CaseStudy2.Controllers;
using CaseStudy2.Service;
using Moq;
using System;
using System.Collections.Generic;
using System.Text;
using Xunit;

namespace AlertToCareTests.ControllerTests
{
    public class MonitorControllerTest
    {
        private Mock<IMonitorService> _mockRepo;

        private MonitorController monitorController;
        public MonitorControllerTest()
        {
            _mockRepo = new Mock<IMonitorService>();
            monitorController = new MonitorController(_mockRepo.Object);
        }

        [Fact]
        public void MonitorRespRatesTest()
        {
            var result = monitorController.MonitorRespRates(1);
            Assert.NotNull(result);
            Assert.IsType<string>(result);
        }
        [Fact]
        public void Monitorspo2sTest()
        {
            var result = monitorController.Monitorspo2s(1);
            Assert.NotNull(result);
            Assert.IsType<string>(result);
        }
        [Fact]
        public void MonitorbpmsTest()
        {
            var result = monitorController.Monitorbpms(1);
            Assert.NotNull(result);
            Assert.IsType<string>(result);
        }
    }
}
